+++ packages/matchers/src/utils/matchPath.ts
@@ -1,6 +1,6 @@
 import * as m from '../../src/matchers'
 import * as t from '@babel/types'
-import { NodePath } from '@babel/traverse'
+import { NodePath } from '@babel/core'
 
 export type CapturedNodePaths<C> = {
   [K in keyof C]: C[K] extends t.Node ? NodePath<C[K]> : C[K]